+++ TauEngine/ios/Example/TauEngine/FLAUTOViewController.h
@@ -1,6 +1,6 @@
 //
 //  FLAUTOViewController.h
-//  flauto_engine
+//  TauEngine
 //
 //  Created by Larpoux on 09/30/2020.
 //  Copyright (c) 2020 Larpoux. All rights reserved.