+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 6.0.1
 
-- Little bug in the Demo App : 58000 is not a valid Sample rate for AAC/ADTS [#460](https://github.com/dooboolab/flutter_sound/issues/460)
+- Little bug in the Demo App : 48000 is not a valid Sample rate for AAC/ADTS [#460](https://github.com/dooboolab/flutter_sound/issues/460)
 
 ## 6.0.0